Rename : 폴더 구조 정리
- model: 데이터 파일
- user_interface: conponents, view
- components : chart, price_info, title, 그 외 컴포넌트 파일들
- view : 화면에 직접 뿌려지는 UI 관련 파일
