+++ README.md
@@ -2,15 +2,16 @@
 
 #### 필수 수행 항목
 
-- [ ] any 를 사용하지 말고 Type을 지정해 주세요
-- [ ] 컴포넌트를 분리하고, 적당한 폴더를 만들어 파일들을 배치하세요
-- [x] 유로화를 입력할 때마다 환율을 반영하여 원화로 표시해 주세요. (exchangeEurToKrw 함수를 사용하세요, 필요시 구현내용 변경 가능)
-- [x] 환율 정보 로딩중임을 사용자가 알 수 있도록 UI를 추가해 주세요.
-- [x] 금액은 세자리 마다 콤마(,) 를 표시하세요
-- [x] 유로화는 소수점 2자리까지 표시하고, 원화는 소수점을 표시하지 않습니다.
-- [ ] 비니지스 로직을 분리하고 MVVM 구현패턴을 적용하세요.
+- [ ] any를 사용하지 말고 Type을 지정
+- [ ] 컴포넌트 분리하고, 적당한 폴더 만들어 파일 배치
+- [x] 유로화 입력할 때마다 환율 반영하여 원화로 표시
+- [x] 환율 정보 로딩중임을 사용자가 알 수 있도록 UI 추가
+- [x] 금액 세자리 마다 콤마(,) 를 표시
+- [x] 유로화 소수점 2자리까지 표시
+- [x] 원화 소수점 표시하지 않음
+- [ ] 비니지스 로직을 분리하고 MVVM 구현패턴을 적용
 
 #### 옵션 수행 항목
 
-- [ ] 비효율적으로 작성된 코드가 있다면 변경하세요
-- [ ] 테스트 코드를 작성하세요
+- [ ] 비효율적으로 작성된 코드가 있다면 변경
+- [ ] 테스트 코드를 작성
