+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import Loader from './Loader';
-import Main from './Main';
+import Loader from './user_interface/components/Loader';
+import Main from './user_interface/view/Main';
 
 export const App = () => {
   const [loading, setLoading] = useState(true);
