+++ src/user_interface/components/CircleCard.tsx
