+++ src/user_interface/components/CountryInfo.tsx
